Change to Part 1
Based on your code in your branch, I revised and finished the part 1 of the project. Note that I create folder called "data" in my notebook and the data files for taxi zones and weather are stored in folders "taxi_zones" and "weather" respectively in "data". If you want to run the code, make sure the files are stored with the same directory.